Mac_Studio/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
ctransformers @ git+https://github.com/marella/ctransformers.git@ed02cf4b9322435972ff3566fd4832806338ca3d
diskcache==5.6.3
fastapi==0.119.1
filelock==3.20.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
llama_cpp_python==0.3.16
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
py-cpuinfo==9.0.0
pydantic==2.12.3
pydantic_core==2.41.4
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
